requirements.txt
pytest==8.3.4
python-dotenv==1.0.1
Requests==2.32.3
setuptools==75.8.0
